requirements.txt
GPUtil==1.4.0
psutil==6.1.1
py-cpuinfo==9.0.0
